+++ src/main/java/insane96mcp/insanesurvivaloverhaul/InsaneSO.java
@@ -21,6 +21,7 @@ import insane96mcp.insanesurvivaloverhaul.module.farming.livestock.LivestockData
 import insane96mcp.insanesurvivaloverhaul.module.farming.plantsgrowth.PlantsGrowthReloadListener;
 import insane96mcp.insanesurvivaloverhaul.module.hungerhealth.HungerAndHealthRegen;
 import insane96mcp.insanesurvivaloverhaul.module.hungerhealth.exhaustion.Exhaustion;
+import insane96mcp.insanesurvivaloverhaul.module.mining.anvilcrafting.AnvilRecipeReloadListener;
 import insane96mcp.insanesurvivaloverhaul.module.mining.blockdefinition.BlockDefinitionReloadListener;
 import insane96mcp.insanesurvivaloverhaul.module.world.CyanFlower;
 import insane96mcp.insanesurvivaloverhaul.setup.ClientSetup;
@@ -112,6 +113,7 @@ public class InsaneSO {
         event.addListener(BlockDefinitionReloadListener.INSTANCE);
         event.addListener(LivestockDataReloadListener.INSTANCE);
         event.addListener(PlantsGrowthReloadListener.INSTANCE);
+        event.addListener(AnvilRecipeReloadListener.INSTANCE);
     }
 
     @SubscribeEvent